backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.17

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0

# Database Migrations
alembic==1.14.0

# Settings Management
pydantic-settings==2.6.1
python-dotenv==1.0.1

python-jose[cryptography]==3.3.0

pandas

celery==5.3.6
redis==5.0.1
aioredis
requests